% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_sparsematrix_rm_weights.R
\name{sparsematrix_rm_weights}
\alias{sparsematrix_rm_weights}
\title{Remove sparse matrix edge weights}
\usage{
sparsematrix_rm_weights(adj_mat)
}
\arguments{
\item{adj_mat}{Sparse matrix of class dgCMatrix}
}
\description{
Removes edge weights from sparse matrices.
}
\examples{
#make matrix
   my_matrix <- sparseMatrix(
       i = c(1, 1, 2, 3, 4, 4, 5, 6, 7, 7), 
       j = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10), 
       x = c(1, 1, 3, 1, 2, 1, 1, 1, 2, 1)
   )
#remove weights
   sparsematrix_rm_weights(my_matrix)
}
\keyword{dgCMatrix}
\keyword{matrix}
